style fixes + error in proforma
