+++ l10n_ca_account_check_writing/__init__.py
@@ -23,5 +23,6 @@
 from . import (
     company,
     account_voucher,
+    account_voucher_line,
     report,
 )
