+++ l10n_ca_account_check_writing/report/l10n_ca_check_print.py
@@ -59,13 +59,10 @@ class report_print_check(report_sxw.rml_parse):
                         'date_due': voucher_line.date_due,
                         'name': name,
                         'amount_original': (
-                            voucher_line.amount_original and
                             voucher_line.amount_original or False),
                         'amount_unreconciled': (
-                            voucher_line.amount_unreconciled and
                             voucher_line.amount_unreconciled or False),
-                        'amount': (voucher_line.amount and
-                                   voucher_line.amount or False),
+                        'amount': (voucher_line.amount or False),
                     }
                     result.append(res)
             else:
@@ -91,13 +88,7 @@ report_sxw.report_sxw(
 report_sxw.report_sxw(
     'report.l10n.ca.account.print.check.middle',
     'account.voucher',
-    'addons/l10n_ca_account_check_writing/report/l10n_ca_check_print_middle.rml',  # NOQA
+    'addons/l10n_ca_account_check_writing/'
+    'report/l10n_ca_check_print_middle.rml',
     parser=report_print_check, header=False
 )
-
-# report_sxw.report_sxw(
-#     'report.l10n.ca.account.print.check.bottom',
-#     'account.voucher',
-#     'addons/l10n_ca_account_check_writing/report/l10n_ca_check_print_bottom.rml',
-#     parser=report_print_check,header=False
-# )
